+++ project-base/tests/ShopBundle/Acceptance/acceptance/ProductImageUploadCest.php
@@ -8,8 +8,8 @@ use Tests\ShopBundle\Test\Codeception\AcceptanceTester;
 
 class ProductImageUploadCest
 {
-    const IMAGE_UPLOAD_FIELD_ID = 'product_edit_form_imageGroup_images_file';
-    const SAVE_BUTTON_NAME = 'product_edit_form[save]';
+    const IMAGE_UPLOAD_FIELD_ID = 'product_form_imageGroup_images_file';
+    const SAVE_BUTTON_NAME = 'product_form[save]';
 
     const EXPECTED_SUCCESS_MESSAGE = 'Product 22" Sencor SLE 22F46DM4 HELLO KITTY modified';